ci: :memo: edit name fooditems
